feat(): fix after review

+++ internal/model/filter.go
@@ -5,6 +5,6 @@ import "time"
 type Filter struct {
 	StartDate time.Time
 	EndDate   time.Time
-	MinPrice  int64
-	MaxPrice  int64
+	MinPrice  float64
+	MaxPrice  float64
 }
