+++ internal/model/price_summary.go
@@ -3,5 +3,5 @@ package model
 type PriceSummary struct {
 	TotalItems      int64
 	TotalCategories int64
-	TotalPrice      int64
+	TotalPrice      float64
 }
